Phase B: ESP32 firmware port (boots on hardware, 129KB free heap)
Port the codec + sync engine to the base CYD (ESP32-2432S028R, no PSRAM),
headless, syncing a .pdb on the SD card to iCloud over Wi-Fi/TLS.

Transport: verified esp_http_client's method enum includes PROPFIND and
REPORT, so dav_esp.c implements the dav.h contract over it (TLS via the
Mozilla cert bundle, Basic auth, chunked decode and redirects handled by the
client) -- more robust than the hand-rolled mbedTLS client the roadmap
tentatively planned. Response parsing is shared with the host: the XML
matchers and the multistatus/report/collection walks moved to dav_xml.c, so
dav.c (curl) and dav_esp.c (mbedtls) parse identically -- the collection-self
filter etc. live in one place.

Storage: PDBs + state on SD (FAT over SPI); sync.c temp paths route under
/sdcard on-device via STATE_DIR.

RAM (the no-PSRAM task): device build shrinks the static arenas behind
ESP_PLATFORM -- MAXR 48, a byte-sized ARENA_CAP (16KB) decoupled from
MAXR*PALM_REC_MAX, 8KB object-fetch buffer, 24KB listing buffer. Boots with
129KB free heap / 110KB largest block; Wi-Fi + SNTP + SD + one
sync_collection wired in app_main. Portability: uint32_t is long on xtensa,
so the UID format sites are cast (host output unchanged).

app_main flow: Wi-Fi STA -> SNTP (no RTC; TLS needs the clock) -> mount SD ->
resolve iCloud effective host -> sync_collection one calendar. Discovery,
category routing, and further streaming are on-device follow-ups. Creds/config
come from firmware/main/secrets.h (gitignored; .example provided).

Host build + all five gates unchanged and green.

+++ bridge/ical.c
@@ -54,7 +54,7 @@ int ical_emit(char *out,int cap,const Appt *a,uint32_t uid){
     textOut(note,sizeof note,a->note);
     const char *tzid = g_tz ? g_tz->id : NULL;
 
-    int n = snprintf(out,cap,"BEGIN:VEVENT\r\nUID:palm-%u@cyd\r\n",uid);
+    int n = snprintf(out,cap,"BEGIN:VEVENT\r\nUID:palm-%u@cyd\r\n",(unsigned)uid);
     if(a->hasTime){
         if(tzid){
             n += snprintf(out+n,cap-n,"DTSTART;TZID=%s:%04d%02d%02dT%02d%02d00\r\n",tzid,a->year,a->month,a->day,a->sH,a->sM);
